backend/requirements.txt
fastapi
uvicorn
feedparser
requests
transformers
torch
vaderSentiment
python-dotenv
sqlalchemy